iox-#1560 Move few classes from helplets to algorithm header
