+++ iceoryx_posh/include/iceoryx_posh/internal/mepoo/typed_mem_pool.hpp
@@ -45,7 +45,7 @@ template <typename T>
 class TypedMemPool
 {
   public:
-    TypedMemPool(const cxx::greater_or_equal<uint32_t, 1> numberOfChunks,
+    TypedMemPool(const algorithm::greater_or_equal<uint32_t, 1> numberOfChunks,
                  BumpAllocator& managementAllocator,
                  BumpAllocator& chunkMemoryAllocator) noexcept;
