+++ iceoryx_posh/source/capro/service_description.cpp
@@ -41,13 +41,13 @@ ServiceDescription::ClassHash::ClassHash(const std::initializer_list<uint32_t>& 
 }
 
 uint32_t&
-ServiceDescription::ClassHash::operator[](iox::cxx::range<uint64_t, 0U, CLASS_HASH_ELEMENT_COUNT - 1> index) noexcept
+ServiceDescription::ClassHash::operator[](iox::algorithm::range<uint64_t, 0U, CLASS_HASH_ELEMENT_COUNT - 1> index) noexcept
 {
     return data[index];
 }
 
 const uint32_t& ServiceDescription::ClassHash::operator[](
-    iox::cxx::range<uint64_t, 0U, CLASS_HASH_ELEMENT_COUNT - 1> index) const noexcept
+    iox::algorithm::range<uint64_t, 0U, CLASS_HASH_ELEMENT_COUNT - 1> index) const noexcept
 {
     return data[index];
 }
